+++ src/main/java/com/insurance/mvp/dto/ReserveWithdrawResponse.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 @Schema(description = "پاسخ عملیات برداشت از اندوخته")
 public class ReserveWithdrawResponse {
     @Schema(description = "کد رایانه", example = "52")
-    private int id;
+    private long id;
 
     @Schema(description = "مبلغ باقیمانده وثیقه", example = "7890000.00")
     private BigDecimal remainedAmount;
@@ -21,5 +21,5 @@ public class ReserveWithdrawResponse {
     private Integer draftId;
 
     @Schema(description = "کد رایانه وثیقه", example = "10")
-    private int collateralId;
+    private long collateralId;
 }
